bring in CeMatchRuleGroup component
